Layer blending (for composite indicators) is now done in one shot instead of a collection/layer at a time.
Moved colorization/legend creation out to a separate set of classes to allow different techniques to be used: SimpleColorizer is the classic, higher-performance version that uses 8 equal-sized bins; QuantileColorizer uses uneven-sized bins based on the complete set of values in a stack of layers, but is slower because it has to fully read all layers.

Framerate can now be specified for animations.