Moved Colorizer and subclasses from layer package into separate color package. Colorizers are now also responsible for colorizing interpreted layers.
Added CustomColorizer which takes a dictionary of tuples of interpreted values (i.e. disturbance type names) to the name of a color palette - can be used for example to group similar disturbance types together with a common color scheme.

Removed all palette arguments and replaced with colorizer instance.

Fixed some layout issues in QuadrantLayout: font sizes, size and position of the scalebar.

Somewhat improved reprojection/cropping in BoundingBox and CompositeIndicator - should now be able to reproject and crop to a bounding box in a projection other than WGS84, which should make per-ha/absolute value conversions quicker when using projections with units in metres.